separate sync command from main typer app
